Provide optimized ASM versions for decode_direct_bits()
We also changed the way the decode_direct_bits() is structured, so that the x86_64 compiler
backend of LLVM can optimize it better. With this version the ASM is only around 1-3% faster.
The aarch64 backend doesn't need this help, since it could already optimize this function
very well. But without this version of the function, the difference between ASM and non ASM
version is 50%. Since LZMA can't use the ASM version at all, we had to make this tradoff
between readability and performance.

We are faster than liblzma in both versions (roughly 20% on x86_64).