Don't overflow dict size when reading LZMA2 streams
There seem to be LZMA2 compressed files out there, that use a dict size of u32::MAX. The original code tried to align to 16 bytes, but we don't need to align anymore.

I would like to remove the need for the function all together, but can't test all edge cases yet.

(And I obviously can't add a test case for this either)